implemented nested fields and split code into multiple files
