+++ pymessagelib/_exceptions.py
@@ -1,0 +1,24 @@
+'''
+Created on Feb 18, 2021
+
+@author: smalb
+'''
+
+class InvalidFieldException(Exception):
+    pass
+
+
+class InvalidFieldDataException(Exception):
+    pass
+
+
+class MissingFieldDataException(Exception):
+    pass
+
+
+class InvalidDataFormatException(Exception):
+    pass
+
+
+class ContextDataMismatchException(Exception):
+    pass
